labscript-agnostic: Generalised batch_compiler so that it can be subclassed to compiler experiments using something other than labscript.
This also involved:
 * creating a new batch compiler for labscript based on the generalised class
 * Moving the globals loading code from labscript to runmanager batch compiler (so that we don't have to duplicate it inside labscript alternatives)
 * Adding version checks for the batch compiler script (now needs labscript 2.4 due to the above point)

There are some additional benefits to this, for instance the __builtins__ dict is saved just before invoking the script, which means that any modules that modify __builtins__ on reload will not have those modifications removed. We also now hold the module watcher and kill lock during cleanup (which I don't think was the case previously unless I don't properly understand when `finally` is executed!)

TODO: Allow the batch_compiler script to be selected by the labconfig file!